adding dev database
It took forever to make this, because of messed up dependences. Thus
it sort of make sense to store this in the repo, but I really couldn't
come down to add it to the master. Adding it to a special branch,
so when we delete it nobody can find this.

Signed-off-by: Zeljko <zeljko@zwr.fi>