Seed database inside transaction
Just making this take a somewhat more reasonable amount of time. It's
just a part of ACID: The database won't tell you it's done with an
INSERT until it actually has it on permanent storage... unless you did
it inside a transaction, in which case it's allowed to cache until
COMMIT and thus can batch writes. I didn't actually test whether the
outcome is the same because it took too long without the
transaction...

Signed-off-by: Ari Rahikkala <ari.rahikkala@gmail.com>